Big commit: config file mode, gui mode
Config file system configuration added, with a parser of the config file that create he system and configure the rooms, devices, and group addresses.
Mode to chose config file and/or gui user interaction added, but the guii cannot interpret user command input and does not represent the system. But the system configured by file can interpret user input through the terminal.